ChatService/pyproject.toml
[tool.poetry]
name = "ChatService"
packages = [{ include = "*", from = "." }]
version = "0.1.0"
description = "基于 FastAPI 的实时聊天微服务"
authors = ["drayee <1473443474@qq.com>"]
readme = "README.md"

exclude = [
    ".venv/**",
    "venv/**",
    ".git/**",
    "__pycache__/**",
    "dist/**",
    "*.log",
    "Dockerfile",
    ".dockerignore"
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.136.0"
uvicorn = "^0.46.0"
python-dotenv = "^1.2.2"
nacos-sdk-python = "^3.2.0b1"
poetry-core = "^2.3.2"
apscheduler = "^3.1.1"
a2a-sdk = "<0.3.23"
asyncpg = "^0.30.0"
redis = "^5.0.0"
httpx = "^0.27.0"
pydantic = "^2.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

